delete and update blog with files

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const BlogSchema = require("../models/blogs.js");
 const jwt = require("jsonwebtoken");
 const path = require("path");
+const fs = require("fs");
 
 const createBlog = async (req, res) => {
   /*  
@@ -12,14 +13,10 @@ const createBlog = async (req, res) => {
     let newBlog;
     let filepath;
     if (req.files) {
-      const files = req.files;
-
-      Object.keys(files).forEach((key) => {
-        filepath = path.join("files", Date.now() + files[key].name);
-        files[key].mv(filepath, (err) => {
-          if (err)
-            return res.status(500).json({ status: "error", message: err });
-        });
+      const files = req.files.file;
+      filepath = path.join("files", Date.now() + files.name);
+      files.mv(filepath, (err) => {
+        if (err) return res.status(500).json({ status: "error", message: err });
       });
       newBlog = await BlogSchema.create({
         title: req.body.title,
@@ -80,17 +77,60 @@ const updateBlog = async (req, res) => {
   /*  
     #swagger.tags = ['Blogs']
 */
+
   try {
+    const token = req.headers.authorization.split(" ")[1];
+    let decodedData = jwt.verify(token, process.env.SECRET_TOKEN);
+
     const { id } = req.params;
-    const updateBlog = await BlogSchema.findByIdAndUpdate(id, req.body, {
-      new: true,
-    });
-    return res.status(200).json({
-      id: updateBlog.id,
-      title: updateBlog.title,
-      description: updateBlog.description,
-      date: updateBlog.date,
-    });
+    const blog = await BlogSchema.findById(id);
+    if (blog.user != decodedData.id) {
+      return res.status(403).json({
+        message: "This Blog is not your",
+      });
+    }
+
+    // const updateBlog = await BlogSchema.findByIdAndUpdate(id, req.body, {
+    //   new: true,
+    // });
+
+    let filepath;
+
+    if (req.files) {
+      const filePath = `${blog.img}`;
+      fs.unlinkSync(filePath);
+
+      const files = req.files.file;
+      filepath = path.join("files", Date.now() + files.name);
+      files.mv(filepath, (err) => {
+        if (err) return res.status(500).json({ status: "error", message: err });
+      });
+      const updateBlog = await BlogSchema.findByIdAndUpdate(
+        id,
+        { ...req.body, img: filepath },
+        {
+          new: true,
+        }
+      );
+      return res.status(200).json({
+        id: updateBlog.id,
+        title: updateBlog.title,
+        description: updateBlog.description,
+        date: updateBlog.date,
+        img: `${req.protocol}://${req.get("host")}/${filepath}`,
+      });
+    } else {
+      const updateBlog = await BlogSchema.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      return res.status(200).json({
+        id: updateBlog.id,
+        title: updateBlog.title,
+        description: updateBlog.description,
+        date: updateBlog.date,
+        img: `${req.protocol}://${req.get("host")}/${blog.img}`,
+      });
+    }
   } catch (error) {
     return res.status(400).json({ message: error.message });
   }
@@ -101,8 +141,20 @@ const deleteBlog = async (req, res) => {
     #swagger.tags = ['Blogs']
 */
   try {
+    const token = req.headers.authorization.split(" ")[1];
+    let decodedData = jwt.verify(token, process.env.SECRET_TOKEN);
+
     const { id } = req.params;
+    const blog = await BlogSchema.findById(id);
+    if (blog.user != decodedData.id) {
+      return res.status(403).json({
+        message: "This Blog is not your",
+      });
+    }
+    const filePath = `${blog.img}`;
+
     const deleteBlog = await BlogSchema.findByIdAndRemove(id);
+    fs.unlinkSync(filePath);
     return res.status(204).json({
       message: "Deleted successfully",
     });
